Add AnalysisNotScheduledSESTM codeunit and SentinelSetup page
This commit adds the AnalysisNotScheduledSESTM codeunit and the SentinelSetup page to the BusinessCentral.Sentinel project. The AnalysisNotScheduledSESTM codeunit is included in the SentinelAdminSESTM permission set, and the SentinelSetup page allows users to configure the Telemetry Logging setting for Sentinel. These changes enhance the functionality and customization options of the application.